WIP replacing separate Playwright process spawning with its JS API
